requirements.txt
pdfplumber
pytesseract
Pillow
python-dotenv
